# Conflicts:
#	README.md
#	bea-content-sync-fusion.php
#	classes/admin/admin-metaboxes.php
#	classes/admin/admin-restrictions.php
#	classes/admin/admin-synchronizations-network.php
#	classes/admin/admin-terms-metaboxes.php
#	classes/admin/admin-terms.php
#	classes/client.php
#	classes/client/attachment.php
#	classes/client/p2p.php
#	classes/client/post_type.php
#	classes/client/taxonomy.php
#	classes/models/synchronization.php
#	classes/models/synchronizations.php
#	classes/plugin.php
#	classes/server/attachment.php
#	classes/server/p2p.php
#	classes/server/post_type.php
#	classes/server/taxonomy.php
#	functions/api.php
#	views/admin/client-terms-form.php
#	views/admin/server-metabox-auto.php
#	views/admin/server-metabox-manual.php
#	views/admin/server-page-add.php
#	views/admin/server-page-settings.php
#	views/admin/server-term-metabox-manual.php

+++ classes/client/taxonomy.php
@@ -1,46 +1,41 @@
 <?php
+
 class BEA_CSF_Client_Taxonomy {
 
 	/**
 	 * Add term on DB
 	 */
-	public static function merge( array $term, BEA_CSF_Synchronization $sync ) {
-		// Test datas validity
-		if ( empty($term) || !is_array( $term ) ) {
+	public static function merge( array $term, array $sync_fields ) {
+		if ( empty( $term ) || ! is_array( $term ) ) {
 			return new WP_Error( 'invalid_datas', __( 'Bad call, invalid datas.' ) );
 		}
-		
-		// Get local ID for parent, from TT_ID
-		$term['parent_tt_id'] = (int) get_term_taxonomy_id_from_meta( '_origin_key', $term['blogid'].':'.(int) $term['parent_tt_id'] );
-		if ( $term['parent_tt_id'] > 0 ) {
-			$term['parent'] = (int) get_term_id_from_term_taxonomy_id( $term['taxonomy'], $term['parent_tt_id'] );
-		}
 
-		// Term exists ?
-		$local_term_id = 0;
-		$local_tt_id = (int) get_term_taxonomy_id_from_meta( '_origin_key', $term['blogid'].':'.(int) $term['term_taxonomy_id'] );
-		if ( $local_tt_id > 0 ) {
-			$local_term_id = (int) get_term_id_from_term_taxonomy_id( $term['taxonomy'], $local_tt_id );
-		}
-
-		if ( $local_term_id == 0 ) {
-			//$edit = false;
-			$new_term_id = wp_insert_term( $term['name'], $term['taxonomy'], array( 'description' => $term['description'], 'slug' => $term['slug'], 'parent' => $term['parent'] ) );
+		$local_term_id = BEA_CSF_Relations::get_object_for_any( 'taxonomy', $term['blogid'], $sync_fields['_current_receiver_blog_id'], $term['term_id'], $term['term_id'] );
+		if ( ! empty( $local_term_id ) && (int) $local_term_id > 0 ) {
+			$new_term_id = wp_update_term( $local_term_id, $term['taxonomy'], array(
+				'name'        => $term['name'],
+				'description' => $term['description'],
+				'slug'        => $term['slug'],
+				'parent'      => $term['parent'],
+			) );
+		} else {
+			$new_term_id = wp_insert_term( $term['name'], $term['taxonomy'], array(
+				'description' => $term['description'],
+				'slug'        => $term['slug'],
+				'parent'      => $term['parent'],
+			) );
 
 			// try to manage error when term already exist with the same name !
 			if ( is_wp_error( $new_term_id ) && $new_term_id->get_error_code() == 'term_exists' ) {
 				$term_exists_result = term_exists( $term['name'], $term['taxonomy'], $term['parent'] );
-				if ( $term_exists_result != false ) {
-					$local_tt_id = (int) get_term_taxonomy_id_from_meta( '_origin_key', $term['blogid'].':'.(int) $term_exists_result['term_taxonomy_id'] );
-					if ( $local_tt_id == 0 ) { // No master ID? no sync item !
+				if ( false != $term_exists_result ) {
+					$local_term_id = BEA_CSF_Relations::get_object_id_for_receiver( 'taxonomy', $sync_fields['_current_receiver_blog_id'], $term['blogid'], (int) $term_exists_result['term_id'] );
+					if ( ! empty( $local_term_id ) && (int) $local_term_id->emitter_id > 0 ) { // No master ID? no sync item !
 						$new_term_id = $term_exists_result;
-						update_term_meta( $term['taxonomy'], $term_exists_result['term_id'], 'already_exists', 1 );
+						update_term_meta( $term_exists_result['term_id'], 'already_exists', 1 );
 					}
 				}
 			}
-		} else {
-			//$edit = true;
-			$new_term_id = wp_update_term( $local_term_id, $term['taxonomy'], array( 'name' => $term['name'], 'description' => $term['description'], 'slug' => $term['slug'], 'parent' => $term['parent'] ) );
 		}
 
 		// Test merge/insertion
@@ -48,69 +43,65 @@ class BEA_CSF_Client_Taxonomy {
 			return new WP_Error( 'term_insertion', $new_term_id->get_error_message() );
 		} elseif ( is_array( $new_term_id ) && isset( $new_term_id['term_id'] ) ) {
 			$new_term_id = (int) $new_term_id['term_id'];
-		} elseif ( (int) $new_term_id != 0 ) {
+		} elseif ( 0 != (int) $new_term_id ) {
 			$new_term_id = (int) $new_term_id;
 		}
 
 		// Always valid ?
-		if ( $new_term_id == 0 ) {
+		if ( 0 === $new_term_id ) {
 			return new WP_Error( 'term_id_invalid', 'Error - Term ID is egal to 0' );
 		}
-		
+
+		// Get term object
+		$new_term_obj = get_term( $new_term_id, $term['taxonomy'] );
+		if ( is_wp_error( $new_term_obj ) ) {
+			return new WP_Error( 'term_valid', 'Error - Term seems invalid' );
+		}
+
+		BEA_CSF_Relations::merge( 'taxonomy', $term['blogid'], $term['term_id'], $GLOBALS['wpdb']->blogid, $new_term_obj->term_id );
+
 		// Save all metas for new post
-		if ( isset( $term['meta_data'] ) && is_array( $term['meta_data'] ) && !empty( $term['meta_data'] ) ) {
+		if ( isset( $term['meta_data'] ) && is_array( $term['meta_data'] ) && ! empty( $term['meta_data'] ) ) {
 			foreach ( $term['meta_data'] as $key => $values ) {
-				// Never sync origin key
-				if ( $key == '_origin_key' ) {
-					continue;
-				}
-				
-				if ( count($values) > 1 || !isset($values[0]) ) {
+				if ( count( $values ) > 1 || ! isset( $values[0] ) ) {
 					// TODO: Management exception, SO RARE in WP !
 					continue;
 				} else {
-					update_term_meta( $term['taxonomy'], $new_term_id, $key, $values[0] );
+					update_term_meta( $new_term_id, $key, $values[0] );
 				}
 			}
 		}
 
-		// Save master id
-		update_term_meta( $term['taxonomy'], $new_term_id, '_origin_key', $term['blogid'].':'.$term['term_taxonomy_id'] );
-
-		// Get full term datas
-		// $new_term = get_term( $new_term_id, $term['taxonomy'] );
-
-		return (int) $new_term_id;
+		return apply_filters( 'bea_csf.client.taxonomy.merge', (int) $new_term_id, $sync_fields );
 	}
 
 	/**
 	 * Delete a term, take the master id, try to find the new ID and delete local term
-	 * 
+	 *
 	 * @param array $term
-	 * @return \WP_Error|boolean
+	 * @param array $sync_fields
+	 *
+	 * @return bool|WP_Error
 	 */
-	public static function delete( array $term, BEA_CSF_Synchronization $sync ) {
-		// Test datas validity
-		if ( empty( $term ) || !is_array( $term ) ) {
+	public static function delete( array $term, array $sync_fields ) {
+		if ( empty( $term ) || ! is_array( $term ) ) {
 			return new WP_Error( 'invalid_datas', __( 'Bad call, invalid datas.' ) );
 		}
 
-		// Term exists ?
-		$local_tt_id = (int) get_term_taxonomy_id_from_meta( '_origin_key', $term['blogid'].':'.(int) $term['term_taxonomy_id'] );
-		if ( $local_tt_id > 0 ) {
+		$local_term_id = BEA_CSF_Relations::get_object_for_any( 'taxonomy', $term['blogid'], $sync_fields['_current_receiver_blog_id'], $term['term_id'], $term['term_id'] );
+		if ( ! empty( $local_term_id ) && (int) $local_term_id > 0 ) {
 			// Term already exist before sync, keep it !
-			$already_exists = (int) get_term_taxonomy_id_from_meta( 'already_exists', 1 );
-			if ( $already_exists == 1 ) {
+			$already_exists = (int) get_term_id_from_meta( 'already_exists', 1 );
+			if ( 1 == $already_exists ) {
 				return false;
 			}
-			
-			$local_term_id = (int) get_term_id_from_term_taxonomy_id( $term['taxonomy'], $local_tt_id );
-			if ( $local_term_id > 0 ) {
-				wp_delete_term( $local_term_id, $term['taxonomy'] );
-			}
+
+			wp_delete_term( $local_term_id, $term['taxonomy'] );
+			BEA_CSF_Relations::delete_by_emitter( 'taxonomy', (int) $GLOBALS['wpdb']->blogid, (int) $local_term_id );
 		}
 
-		return true;
-	}
+		BEA_CSF_Relations::delete_by_emitter( 'taxonomy', (int) $term['blogid'], (int) $term['term_id'] );
 
+		return apply_filters( 'bea_csf.client.taxonomy.delete', true, $sync_fields );
+	}
 }
